+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlayground.swift
@@ -27,6 +27,7 @@ struct PaymentSheetTestPlayground: View {
         SettingView(setting: $playgroundController.settings.defaultBillingAddress)
         SettingView(setting: $playgroundController.settings.linkEnabled)
         SettingView(setting: $playgroundController.settings.externalPayPalEnabled)
+        SettingView(setting: $playgroundController.settings.preferredNetworksEnabled)
         SettingView(setting: $playgroundController.settings.autoreload)
     }