+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## X.X.X
 ### PaymentsUI
 * [Fixed] An issue with `STPPaymentCardTextField`, where the card params were not updated after deleting an empty sub field.
+* [Fixed] Switched to Asset Catalogs and updated to the latest card brand logos.
 
 ## 23.17.1 2023-10-09
 ### PaymentSheet